+++ controller_interface/include/controller_interface/controller_interface_params.hpp
@@ -47,6 +47,7 @@ struct ControllerInterfaceParams
   std::string controller_name = "";
   std::string robot_description = "";
   unsigned int update_rate = 0;
+  unsigned int controller_manager_update_rate = 0;
 
   std::string node_namespace = "";
   rclcpp::NodeOptions node_options = rclcpp::NodeOptions();